python-app-test/requirements.txt
falcon
gunicorn
requests
gevent
msgpack
fluent-logger